+++ nixjs-rt/src/builtins.ts
@@ -1,7 +1,7 @@
-import { err } from ".";
-import { abortError } from "./abort";
-import { otherError } from "./other";
-import { typeMismatchError } from "./typeError";
+import { err } from "./errors";
+import { abortError } from "./errors/abort";
+import { otherError } from "./errors/other";
+import { typeMismatchError } from "./errors/typeError";
 import {
   Attrset,
   EvalCtx,
@@ -16,7 +16,7 @@ import {
   NixType,
   Path,
   TRUE,
-} from "../lib";
+} from "./lib";
 
 type BuiltinsRecord = Record<string, (param: NixType) => NixType>;
